Initial commit
This is the first version of the script. It scans your inbox for Red Bubble Sales messages, sorts them into a folder for you and collates the data into a table. I made this in Pycharm with Python 3. Your gmail account will need to be set up for IMAP for this to work. Feel free to use or modify as you will.